// @components/types/MatchDetails.tsx
export interface MatchDetails {
    team: string;
    goals: any[]; // You can replace 'any' with a more specific type if needed, like 'number[]' for goals.
    ycards: any[]; // Same as goals, replace 'any' with a more specific type if necessary.
    rcards: any[]; // Same as goals, replace 'any' with a more specific type if necessary.
}